Add ICD-10-CM Chapter Labels

Add ICD-10-CM Chapter Labels

Usage

case_chapter_icd10(df, col)

Arguments

df

<data.frame>data frame

col

<sym> unquoted column name of ICD-10-CM codes to match

Value

A tibble with a chapter column

Examples

dplyr::tibble(code = c("F50.8", "G40.311", "Q96.8", "Z62.890", "R45.4",
"E06.3", "H00.019", "D50.1", "C4A.70", "Z20.818")) |>
case_chapter_icd10(code)
#> # A tibble: 10 × 2
#>    code    chapter                                                              
#>    <chr>   <chr>                                                                
#>  1 F50.8   Mental, behavioral and neurodevelopmental disorders                  
#>  2 G40.311 Diseases of the nervous system                                       
#>  3 Q96.8   Congenital malformations, deformations and chromosomal abnormalities 
#>  4 Z62.890 Factors influencing health status and contact with health services   
#>  5 R45.4   Symptoms, signs and abnormal clinical and laboratory findings, not e…
#>  6 E06.3   Endocrine, nutritional and metabolic diseases                        
#>  7 H00.019 Diseases of the eye and adnexa                                       
#>  8 D50.1   Diseases of the blood and blood-forming organs and certain disorders…
#>  9 C4A.70  Neoplasms                                                            
#> 10 Z20.818 Factors influencing health status and contact with health services