Match ICD-10-CM Codes to Chapter Labels

Match ICD-10-CM Codes to Chapter Labels

Usage

icd_code_to_chapter(x)

Arguments

x

<chr> vector of ICD-10-CM codes

Value

<chr> vector of ICD-10-CM chapter labels

Examples

codes <- c("F50.8", "G40.311", "Q96.8",
           "R45.4", "E06.3", "H00.019",
           "D50.1", "C4A.70", "Z20.818")

icd_code_to_chapter(codes)
#> [1] "Mental, behavioral and neurodevelopmental disorders"                                                
#> [2] "Diseases of the nervous system"                                                                     
#> [3] "Congenital malformations, deformations and chromosomal abnormalities"                               
#> [4] "Symptoms, signs and abnormal clinical and laboratory findings, not elsewhere classified"            
#> [5] "Endocrine, nutritional and metabolic diseases"                                                      
#> [6] "Diseases of the eye and adnexa"                                                                     
#> [7] "Diseases of the blood and blood-forming organs and certain disorders involving the immune mechanism"
#> [8] "Neoplasms"                                                                                          
#> [9] "Factors influencing health status and contact with health services"                                 

data.frame(code = codes,
           chapter = icd_code_to_chapter(codes))
#>      code
#> 1   F50.8
#> 2 G40.311
#> 3   Q96.8
#> 4   R45.4
#> 5   E06.3
#> 6 H00.019
#> 7   D50.1
#> 8  C4A.70
#> 9 Z20.818
#>                                                                                               chapter
#> 1                                                 Mental, behavioral and neurodevelopmental disorders
#> 2                                                                      Diseases of the nervous system
#> 3                                Congenital malformations, deformations and chromosomal abnormalities
#> 4             Symptoms, signs and abnormal clinical and laboratory findings, not elsewhere classified
#> 5                                                       Endocrine, nutritional and metabolic diseases
#> 6                                                                      Diseases of the eye and adnexa
#> 7 Diseases of the blood and blood-forming organs and certain disorders involving the immune mechanism
#> 8                                                                                           Neoplasms
#> 9                                  Factors influencing health status and contact with health services