ICD-10-CM Code Edits

Definitions of Medicare Code Edits, version 41.1

Usage

search_edits(icd = NULL)

Arguments

icd

<chr> vector of ICD-10-CM codes

Value

a tibble

Details

FY 2024 – Version 41.1 (Effective April 1, 2024 through September 30, 2024)

The ICD-10 Definitions of Medicare Code Edits file contains the following: A description of each coding edit with the corresponding code lists as well as all the edits and the code lists effective for FY 2024.

Examples

search_edits(c("Q96.8", "N47.0", "R45.4", "A33"))
#> # A tibble: 4 × 4
#>   icd_code icd_description                  icd_conflict_group icd_conflict_rule
#>   <chr>    <chr>                            <chr>              <chr>            
#> 1 A33      Tetanus neonatorum               Age                Perinatal/Newbor…
#> 2 N47.0    Adherent prepuce, newborn        Age                Perinatal/Newbor…
#> 3 Q96.8    Other variants of Turner's synd… Sex                Female Only      
#> 4 N47.0    Adherent prepuce, newborn        Sex                Male Only